+++ src/RouteGenerator/RequestAwareRouteGeneratorFactory.php
@@ -6,17 +6,24 @@ use Pagerfanta\RouteGenerator\RouteGeneratorFactoryInterface;
 use Pagerfanta\RouteGenerator\RouteGeneratorInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
+use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
 final class RequestAwareRouteGeneratorFactory implements RouteGeneratorFactoryInterface
 {
     private UrlGeneratorInterface $router;
     private RequestStack $requestStack;
+    private ?PropertyAccessorInterface $propertyAccessor;
 
-    public function __construct(UrlGeneratorInterface $router, RequestStack $requestStack)
+    public function __construct(UrlGeneratorInterface $router, RequestStack $requestStack, ?PropertyAccessorInterface $propertyAccessor = null)
     {
+        if (null === $propertyAccessor) {
+            trigger_deprecation('babdev/pagerfanta-bundle', '3.1', 'Not passing a "%s" to the "%s" constructor is deprecated. As of 4.0, it will be a required argument.', PropertyAccessorInterface::class, self::class);
+        }
+
         $this->router = $router;
         $this->requestStack = $requestStack;
+        $this->propertyAccessor = $propertyAccessor;
     }
 
     public function create(array $options = []): RouteGeneratorInterface
@@ -46,7 +53,11 @@ final class RequestAwareRouteGeneratorFactory implements RouteGeneratorFactoryIn
             $options['routeParams'] = array_merge($defaultRouteParams, $options['routeParams']);
         }
 
-        return new RouterAwareRouteGenerator($this->router, $options);
+        return new RouterAwareRouteGenerator(
+            $this->router,
+            null === $this->propertyAccessor ? $options : $this->propertyAccessor,
+            null === $this->propertyAccessor ? [] : $options
+        );
     }
 
     private function getRequest(): ?Request